+++ src/utils/duplicateReducer.js
@@ -1,6 +1,6 @@
-module.exports.duplicateReducer = (acc, value, key, arr) =>
-  acc.concat(
-    arr.filter(x => x === value).length > 1 && acc.indexOf(value) === -1
-      ? value
-      : []
-  );
+module.exports.duplicateReducer = (acc, value, key, arr) => [
+  ...acc,
+  ...(arr.filter(x => x === value).length > 1 && acc.indexOf(value) === -1
+    ? value
+    : [])
+];